Keep vext .so files with -p debug=+vmod_so_keep
Required for seamless debugging.

Alternatives considered:
- add vext_so_keep debug flag -> overkill

To decide:
- rename to libso_keep / "Keep copied VMOD/VEXT libraries" ?
  (libso to disambiguate from so for socket option)

+++ bin/varnishd/common/common_vext.c
@@ -158,14 +158,15 @@ vext_load(void)
 }
 
 void
-vext_cleanup(void)
+vext_cleanup(int do_unlink)
 {
 	struct vext *vp;
 
 	VTAILQ_FOREACH(vp, &vext_list, list) {
 		fprintf(stderr, "ee3 %s\n", VSB_data(vp->vsb));
 		if (vp->vsb != NULL && VSB_len(vp->vsb) > 0) {
-			XXXAZ(unlink(VSB_data(vp->vsb)));
+			if (do_unlink)
+				XXXAZ(unlink(VSB_data(vp->vsb)));
 			VSB_clear(vp->vsb);
 		}
 	}
